Method Statement was refactored.

+++ src/user_parameters.h
@@ -8,7 +8,7 @@ struct AccountInfo{
 
   string login_ = "no_name";
   string password_ = "0000";
-  double cash_ = 0.0;
+  int cash_ = 0.0;
   double credit_ = 0.0;
   double monthly_payment_ = 0.0;
   int amount_of_credit_month_ = 0;
